fix one minor import in stu's work relative to changes;rename dir
